requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-socketio>=5.13.0
python-engineio>=4.12.2
psycopg2-binary==2.9.9
requests>=2.32.2
pygame==2.5.2
Werkzeug==3.0.3
gunicorn==22.0.0
locust==2.29.0